lol move non-test files out of test directories
