+++ frontend/lib/pages/fee-waiver.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 
-import { FeeWaiverMiscInput, FeeWaiverIncomeInput, FeeWaiverExpensesInput } from "../../queries/globalTypes";
-import Page from "../../page";
-import { SessionUpdatingFormSubmitter } from '../../forms';
-import Routes from '../../routes';
-import { CheckboxFormField } from '../../form-fields';
-import { YesNoRadiosFormField } from "../../yes-no-radios-form-field";
-import { BackButton, NextButton } from '../../buttons';
-import { CurrencyFormField } from '../../currency-form-field';
-import { getInitialFormInput } from '../../form-input-converter';
-import { FeeWaiverMiscMutation } from '../../queries/FeeWaiverMiscMutation';
-import { FeeWaiverIncomeMutation } from '../../queries/FeeWaiverIncomeMutation';
-import { FeeWaiverExpensesMutation } from '../../queries/FeeWaiverExpensesMutation';
+import { FeeWaiverMiscInput, FeeWaiverIncomeInput, FeeWaiverExpensesInput } from "../queries/globalTypes";
+import Page from "../page";
+import { SessionUpdatingFormSubmitter } from '../forms';
+import Routes from '../routes';
+import { CheckboxFormField } from '../form-fields';
+import { YesNoRadiosFormField } from "../yes-no-radios-form-field";
+import { BackButton, NextButton } from '../buttons';
+import { CurrencyFormField } from '../currency-form-field';
+import { getInitialFormInput } from '../form-input-converter';
+import { FeeWaiverMiscMutation } from '../queries/FeeWaiverMiscMutation';
+import { FeeWaiverIncomeMutation } from '../queries/FeeWaiverIncomeMutation';
+import { FeeWaiverExpensesMutation } from '../queries/FeeWaiverExpensesMutation';
 import { ProgressiveOtherCheckboxFormField } from '../other-checkbox-form-field';
 
 const INITIAL_MISC_STATE: FeeWaiverMiscInput = {
